Add BPEL activities to the model.
To parse the incoming bpel file there has to be an internal representation of the BPEL model. In this commit all the activities are (partly) added to that model. The same structure is used as in the official WS-BPEL specification. For activities that is: basic and structured. The standard attributes and elements for all activities are extracted to a superclass Activities.